+++ HISTORY.md
@@ -1,5 +1,8 @@
 History of features added to LiChess Tools in time (reversed order)
 
+09 Aug
+  - ceval line options - highlight
+
 06 Aug
   - LT engine configurable threads and hash
   - Dim pieces when out of main line